Removed SigninTest dependencies from BlackBoxBase and BBBUtil and moved methods over to BBBUtil
