+++ orcid-integration-test/src/test/java/org/orcid/integration/blackbox/web/SigninTest.java
@@ -16,18 +16,9 @@
  */
 package org.orcid.integration.blackbox.web;
 
-import java.util.List;
-
 import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.openqa.selenium.By;
-import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
-import org.orcid.integration.blackbox.api.BBBUtil;
 import org.orcid.integration.blackbox.api.BlackBoxBase;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.test.context.ContextConfiguration;
@@ -56,7 +47,7 @@ public class SigninTest extends BlackBoxBase {
         webDriver.get(baseUri + "/userStatus.json?logUserOut=true");
         webDriver.get(baseUri + "/my-orcid");
         signIn(webDriver, user1UserName, user1Password);
-        dismissVerifyEmailModal(webDriver);
+        dismissVerifyEmailModal();
     }
     
     @Test
@@ -73,40 +64,7 @@ public class SigninTest extends BlackBoxBase {
     	webDriver.get(baseUri + "/userStatus.json?logUserOut=true");
         webDriver.get(baseUri + "/my-orcid");
         signIn(webDriver, userName, user1Password);
-        dismissVerifyEmailModal(webDriver);
-    }
-
-	// Make this available to other classes
-    static public void signIn(WebDriver webDriver, String username, String password) {
-        BBBUtil.extremeWaitFor(BBBUtil.documentReady(),webDriver);
-        BBBUtil.extremeWaitFor(BBBUtil.angularHasFinishedProcessing(),webDriver);
-        BBBUtil.extremeWaitFor(ExpectedConditions.presenceOfElementLocated(By.xpath("//input[@name='userId']")), webDriver);
-        BBBUtil.extremeWaitFor(BBBUtil.angularHasFinishedProcessing(),webDriver);
-        WebElement emailEl = webDriver.findElement(By.xpath("//input[@name='userId']"));
-        emailEl.sendKeys(username);
-        BBBUtil.extremeWaitFor(BBBUtil.angularHasFinishedProcessing(),webDriver);
-        WebElement passwordEl = webDriver.findElement(By.xpath("//input[@name='password']"));
-        passwordEl.sendKeys(password);
-        BBBUtil.extremeWaitFor(BBBUtil.angularHasFinishedProcessing(),webDriver);
-        WebElement buttonEl = webDriver.findElement(By.xpath("//button[@id='form-sign-in-button']"));
-        BBBUtil.ngAwareClick(buttonEl, webDriver);
-        BBBUtil.extremeWaitFor(ExpectedConditions.presenceOfElementLocated(By.xpath("//a[text()='Sign out']")),webDriver);
-        BBBUtil.extremeWaitFor(BBBUtil.documentReady(),webDriver);
-        BBBUtil.extremeWaitFor(BBBUtil.angularHasFinishedProcessing(),webDriver);
-    }
-
-    public static void dismissVerifyEmailModal(WebDriver webDriver) {
-        WebDriverWait wait = new WebDriverWait(webDriver, 10);
-        List<WebElement> weList = webDriver.findElements(By.xpath("//div[@ng-controller='VerifyEmailCtrl']"));
-        if (weList.size() > 0) {// we need to wait for the color box to appear
-            wait.until(ExpectedConditions.presenceOfAllElementsLocatedBy(By.xpath("//div[@ng-controller='VerifyEmailCtrl' and @orcid-loading='false']")));
-            ((JavascriptExecutor) webDriver).executeScript("$.colorbox.close();");
-            colorBoxIsClosed(wait);
-        }
-    }
-
-    public static void colorBoxIsClosed(WebDriverWait wait) {
-        wait.until(ExpectedConditions.not(ExpectedConditions.visibilityOfAllElementsLocatedBy(By.xpath("//div[@id='colorbox']"))));
+        dismissVerifyEmailModal();
     }
 
 }
