+++ orcid-integration-test/src/test/java/org/orcid/integration/blackbox/web/shibboleth/ShibbolethTest.java
@@ -33,7 +33,7 @@ import org.openqa.selenium.firefox.FirefoxProfile;
 import org.openqa.selenium.remote.DesiredCapabilities;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import org.orcid.integration.blackbox.web.SigninTest;
+import org.orcid.integration.blackbox.api.BlackBoxBase;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
@@ -98,7 +98,7 @@ public class ShibbolethTest {
         webDriver.get(baseUri + "/userStatus.json?logUserOut=true");
         webDriver.get(baseUri + "/shibboleth/signin");
         new WebDriverWait(webDriver, DEFAULT_TIMEOUT_SECONDS).until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//input[@name='userId']")));
-        SigninTest.signIn(webDriver, user1UserName, user1Password);
+        BlackBoxBase.signIn(webDriver, user1UserName, user1Password);
         new WebDriverWait(webDriver, DEFAULT_TIMEOUT_SECONDS).until(ExpectedConditions.presenceOfElementLocated(By.xpath("//span[contains(., '" + user1OrcidId + "')]")));
         // Check can sign in again without linking
         webDriver.get(baseUri + "/userStatus.json?logUserOut=true");
